- GPU-driven instancing is now fully implemented - Fixed a bug where subsequent draw instanced calls would use instance data from index 0 instead of the correct offset - Small refactor of DX12 shader reflection - Bugfixing

+++ Engine/Source/Rendering/DX12/CompactSMInstancesComputeShader.cpp
@@ -1,5 +1,4 @@
 ﻿#include "CompactSMInstancesComputeShader.h"
-
 #include "DX12RenderingSubsystem.h"
 
 CompactSMInstancesComputeShader::CompactSMInstancesComputeShader(const CompactSMInstancesComputeShader& other) :
@@ -24,7 +23,7 @@ void CompactSMInstancesComputeShader::Run(DX12GraphicsCommandList& commandList,
 
     _instanceBuffer = &instanceBuffer;
     _rootConstants.InstanceCount = instanceCount;
-    _rootConstants.InstanceCapacity = instanceBuffer.GetData().Capacity();
+    _rootConstants.InstanceCapacity = static_cast<uint32>(instanceBuffer.GetData().Capacity());
     
     const uint32 threads = std::max(instanceCount / 32u, 1u);
     Dispatch(commandList, threads, 1, 1);
